+++ src/utils/timing.rs
@@ -6,11 +6,11 @@ use tracing::info;
 
 use crate::utils::text::truncate_to_chars;
 
-pub fn now_unix_seconds() -> u64 {
+pub fn now_unix_seconds() -> i64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default()
-        .as_secs()
+        .as_secs() as i64
 }
 
 #[derive(Debug)]
